move ModuleOptions into base with other base classes
