+++ runway/module/cdk.py
@@ -15,7 +15,7 @@ from .base import RunwayModuleNpm
 if TYPE_CHECKING:
     from .._logging import RunwayLogger
     from ..context.runway import RunwayContext
-    from . import ModuleOptions
+    from .base import ModuleOptions
 
 LOGGER = cast("RunwayLogger", logging.getLogger(__name__))
