+++ runway/module/k8s.py
@@ -17,7 +17,7 @@ from .base import RunwayModule
 if TYPE_CHECKING:
     from .._logging import RunwayLogger
     from ..context.runway import RunwayContext
-    from . import ModuleOptions
+    from .base import ModuleOptions
 
 LOGGER = cast("RunwayLogger", logging.getLogger(__name__))
