+++ tests/unit/module/test_base.py
@@ -3,14 +3,16 @@
 from __future__ import annotations
 
 import logging
+from contextlib import contextmanager
 from pathlib import Path
-from typing import TYPE_CHECKING, Any, Dict, List
+from typing import TYPE_CHECKING, Any, Dict, Iterator, List
 
 import pytest
 
 from runway.exceptions import NpmNotFound
 from runway.module.base import (
     NPM_BIN,
+    ModuleOptions,
     RunwayModule,
     RunwayModuleNpm,
     format_npm_command_for_logging,
@@ -25,6 +27,12 @@ if TYPE_CHECKING:
     from ..factories import MockRunwayContext
 
 MODULE = "runway.module.base"
+
+
+@contextmanager
+def does_not_raise() -> Iterator[None]:
+    """Use for conditional pytest.raises when using parametrize."""
+    yield
 
 
 @pytest.mark.parametrize(
@@ -93,6 +101,66 @@ def test_use_npm_ci(
         assert fake_process.call_count(cmd) == 1
     else:
         assert fake_process.call_count(cmd) == 0
+
+
+class TestModuleOptions:
+    """Test runway.module.ModuleOptions."""
+
+    @pytest.mark.parametrize(
+        "data, env, expected, exception",
+        [
+            ("test", None, "test", does_not_raise()),
+            ("test", "env", "test", does_not_raise()),
+            ({"key": "value"}, "test", {"key": "called"}, does_not_raise()),
+            ({"key": "value"}, None, {"key": "called"}, does_not_raise()),
+            (["test"], None, ["test"], does_not_raise()),
+            (["test"], "env", ["test"], does_not_raise()),
+            (None, None, None, does_not_raise()),
+            (None, "env", None, does_not_raise()),
+            (100, None, None, pytest.raises(TypeError)),
+            (100, "env", None, pytest.raises(TypeError)),
+        ],
+    )
+    def test_merge_nested_env_dicts(
+        self, mocker: MockerFixture, data: Any, env: Any, expected: Any, exception: Any
+    ) -> None:
+        """Test merge_nested_env_dicts."""
+
+        def merge_nested_environment_dicts(value, env_name):
+            """Assert args passed to the method during parse."""
+            assert value == list(data.values())[0]  # only pass data dict of 1 item
+            assert env_name == env
+            return "called"
+
+        mocker.patch(
+            f"{MODULE}.merge_nested_environment_dicts", merge_nested_environment_dicts,
+        )
+
+        with exception:
+            assert ModuleOptions.merge_nested_env_dicts(data, env) == expected
+
+    def test_parse(self, runway_context: MockRunwayContext) -> None:
+        """Test parse."""
+        with pytest.raises(NotImplementedError):
+            assert ModuleOptions.parse(runway_context)
+
+    @pytest.mark.parametrize("value", ["test", None, 123, {"key": "val"}])
+    def test_mutablemapping_abstractmethods(self, value: Any) -> None:
+        """Test the abstractmethods of MutableMapping."""
+        obj = ModuleOptions()
+        obj["key"] = value  # __setitem__
+        assert obj["key"] == value, "test __setitem__ and __getitem__"
+        assert isinstance(obj["key"], type(value)), "test type retention"
+        assert obj.get("key") == value, "ensure .get() is working"
+        assert len(obj) == 1, "test __len__"
+
+        counter = sum(1 for _ in obj)
+        assert counter == 1, "test __iter__"
+
+        del obj["key"]
+        assert "key" not in obj, "test __delitem__ with __contains__"
+        with pytest.raises(KeyError):
+            assert obj["key"], "test __delitem__ with __getitem__"
 
 
 class TestRunwayModuleNpm:
